# models/Zuct_db/generated/StudentsModelGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE studentsModelGenerated.js PLEASE EDIT ../studentsModel.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
// Database
import Database from "../../../classes/Database_Zuct_db";
import mongoose, { Schema } from "mongoose";

// Logger
import Logger from "../../../classes/Logger";

const generatedModel = {
  /**
   * Init  schema
   */
  init() {
    const db = Database.getConnection();

    /**
      * students
      */
    const studentsSchema = new mongoose.Schema({
      date_of_birth: {
        type: "Date", 
        required: true
      },
      first_name: {
        type: "String", 
        required: true
      },
      gender: {
        type: "String", 
        required: true
      },
      last_name: {
        type: "String", 
        required: true
      },
      middle_name: {
        type: "String"
      },
      phone: {
        type: "Number"
      },
      // RELATIONS
      _enroll: {
        type: Schema.ObjectId,
        ref: "programs"
      },
      _takes: [{
        type: Schema.ObjectId,
        ref: "courses"
      }],
      _taught: [{
        type: Schema.ObjectId,
        ref: "lecturers"
      }],
      
      
      // EXTERNAL RELATIONS
      /*
      */
    });

    generatedModel.setModel(db.connection.model("Students", studentsSchema));

    return studentsSchema;
  },

  /**
   * Set Model
   */
  setModel: model => {
    generatedModel.model = model;
  },

  /**
   * Get model
   */
  getModel: () => {
    return generatedModel.model;
  },

  // Start queries
    

  // CRUD METHODS


  /**
  * studentsModel.create
  *   @description CRUD ACTION create
  *
  */
  async create(item) {
    const obj = new generatedModel.model(item);
    return await obj.save();
  },
  
  /**
  * studentsModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  async delete(id) {
    return await generatedModel.model.findByIdAndRemove(id);
  },
  
  /**
  * studentsModel.findBy_enroll
  *   @description CRUD ACTION findBy_enroll
  *   @param Objectid key Id of model to search for
  *
  */
  async findBy_enroll(key) {
    return await generatedModel.model.find({ '_enroll' : key});
  },
  
  /**
  * studentsModel.findBy_takes
  *   @description CRUD ACTION findBy_takes
  *   @param Objectid key Id of model to search for
  *
  */
  async findBy_takes(key) {
    return await generatedModel.model.find({ '_takes' : key});
  },
  
  /**
  * studentsModel.findBy_taught
  *   @description CRUD ACTION findBy_taught
  *   @param Objectid key Id of model to search for
  *
  */
  async findBy_taught(key) {
    return await generatedModel.model.find({ '_taught' : key});
  },
  
  /**
  * studentsModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *
  */
  async get(id) {
    return await generatedModel.model.findOne({ _id : id });
  },
  
  /**
  * studentsModel.list
  *   @description CRUD ACTION list
  *
  */
  async list() {
    return await generatedModel.model.find();
  },
  
  /**
  * studentsModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *
  */
  async update(item) { 
    return await generatedModel.model.findOneAndUpdate({ _id: item._id }, item, {'new': true});
  },
  


};

export default generatedModel;